refactored Error handling to Try
