+++ src/main/scala/com/neu/Qualitycheck/QualityCheckResult.scala
@@ -1,0 +1,7 @@
+package com.neu.Qualitycheck
+
+case class QualityCheckResult(
+    checkName: String,
+    passed: Boolean,
+    failedRecords: Option[Seq[String]] = None
+)
