--- src/main/scala/com/neu/qualitycheck/QualityCheckResult.scala
@@ -1,7 +1,0 @@
-package com.neu.qualitycheck
-
-case class QualityCheckResult(
-    checkName: String,
-    passed: Boolean,
-    failedRecords: Option[Seq[String]] = None
-)
